# docs/swagger.json
{
    "components": {
        "schemas": {
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "validation failed",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "model.MatchCriteria": {
                "properties": {
                    "injection_point": {
                        "description": "ingress/egress/transient/custom, \"\" = any",
                        "type": "string"
                    },
                    "labels": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "description": "AND semantics",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "model.Rule": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "fault_composition_id": {
                        "type": "string"
                    },
                    "fault_spec_id": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "match": {
                        "$ref": "#/components/schemas/model.MatchCriteria"
                    },
                    "mode": {
                        "description": "\"inline\" or \"background\"",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "integer"
                    },
                    "service": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "Central coordination controller for the atropos ecosystem.\nManages rules, faults, experiments, workflows, and SDK lifecycle.",
        "title": "Manteion Control-Plane API",
        "version": "1.0"
    },
    "paths": {
...
        "/rules": {
            "get": {
                "description": "Returns all configured rules, ordered by priority descending.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/model.Rule"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "internal error"
                    }
                },
                "summary": "List rules",
                "tags": [
                    "rules"
                ]
            },
            "post": {
                "description": "Persist a new rule. The store layer validates the rule per\nmodel.Rule.Validate() (id/name/service required, exactly one of\nfault_spec_id or fault_composition_id, mode in {inline,background}).",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/model.Rule",
                                        "summary": "rule",
                                        "description": "Rule definition"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Rule definition",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.Rule"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "validation error or infrastructure failure (currently collapsed; see openapi-conventions.md known gaps)"
                    }
                },
                "summary": "Create rule",
                "tags": [
                    "rules"
                ]
            }
        },
        "/rules/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "Rule ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "rule deleted"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "rule not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Delete rule",
                "tags": [
                    "rules"
                ]
            },
            "get": {
                "parameters": [
                    {
                        "description": "Rule ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.Rule"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "rule not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Get rule",
                "tags": [
                    "rules"
                ]
            },
            "put": {
                "description": "Replace an existing rule. The id in the path overrides any id in the body.",
                "parameters": [
                    {
                        "description": "Rule ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/model.Rule",
                                        "summary": "rule",
                                        "description": "Updated rule"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Updated rule",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.Rule"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "invalid JSON or validation error"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "rule not found"
                    }
                },
                "summary": "Update rule",
                "tags": [
                    "rules"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Local dev (HTTP)",
            "url": "http://localhost:8080/api/v1"
        },
        {
            "description": "Local dev (HTTPS)",
            "url": "https://localhost:8080/api/v1"
        }
    ]
}